nodes/koi-net-processor-b-node/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "processor-b-node"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "httpx", 
    "python-dotenv",
    "rid-lib>=3.2.3", # Ensure compatible versions
    "koi-net==1.0.0b12", 
    "rich", 
    "ruamel.yaml",
    # Add markdown parser if needed, e.g.:
    # "markdown-it-py", 
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["processor_b_node"] # Specify the package 